# server/docs/swagger.yaml
basePath: /
definitions:
...
  response.PageResp:
    properties:
      count:
        description: 总数
        type: integer
      lists:
        description: 数据列表
      pageNo:
        description: 当前页码
        type: integer
      pageSize:
        description: 每页数量
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  schema.UserProtocolResp:
    properties:
      content:
        description: 协议内容
        type: string
      createTime:
        description: 创建时间
        type: string
      createdBy:
        description: 创建人id
        type: string
      createdByUser:
        description: 创建人
        type: string
      id:
        type: string
      tag:
        description: 标识
        type: string
      title:
        description: 标题
        type: string
      updateTime:
        description: 更新时间
        type: string
      version:
        description: 版本
        type: number
    type: object
...
info:
  contact:
    email: x@adtk.cn
    name: xh
    url: http://x.adtk.cn
  description: x_admin后台管理系统
  license:
    name: MIT License
    url: https://gitee.com/xiangheng/x_admin/blob/main/LICENSE
  termsOfService: http://x.adtk.cn
paths:
...
  /api/admin/user_protocol/add:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 标识
        in: body
        name: Tag
        schema:
          type: string
      - description: 版本
        in: body
        name: Version
        schema:
          type: number
      - description: 标题
        in: body
        name: Title
        schema:
          type: string
      - description: 协议内容
        in: body
        name: Content
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户协议新增
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/del:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Id
        in: body
        name: Id
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户协议删除
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/del_batch:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 逗号分割的id
        in: body
        name: Ids
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户协议删除-批量
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/detail:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Id
        in: query
        name: Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/schema.UserProtocolResp'
              type: object
      summary: 用户协议详情
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/edit:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Id
        in: body
        name: Id
        schema:
          type: string
      - description: 标识
        in: body
        name: Tag
        schema:
          type: string
      - description: 版本
        in: body
        name: Version
        schema:
          type: number
      - description: 标题
        in: body
        name: Title
        schema:
          type: string
      - description: 协议内容
        in: body
        name: Content
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户协议编辑
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/export_file:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 标题
        in: query
        name: Title
        type: string
      - description: 协议内容
        in: query
        name: Content
        type: string
      - description: 版本
        in: query
        name: Version
        type: number
      - description: 创建时间
        in: query
        name: CreateTimeStart
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeEnd
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeStart
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeEnd
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: file
        "500":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户协议导出
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/import_file:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 导入文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户协议导入
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/list:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 页码
        in: query
        name: pageNo
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 标题
        in: query
        name: Title
        type: string
      - description: 协议内容
        in: query
        name: Content
        type: string
      - description: 版本
        in: query
        name: Version
        type: number
      - description: 创建时间
        in: query
        name: CreateTimeStart
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeEnd
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeStart
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeEnd
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PageResp'
                  - properties:
                      lists:
                        items:
                          $ref: '#/definitions/schema.UserProtocolResp'
                        type: array
                    type: object
              type: object
      summary: 用户协议列表
      tags:
      - user_protocol-用户协议
  /api/admin/user_protocol/list_all:
    get:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: 标题
        in: query
        name: Title
        type: string
      - description: 协议内容
        in: query
        name: Content
        type: string
      - description: 版本
        in: query
        name: Version
        type: number
      - description: 创建时间
        in: query
        name: CreateTimeStart
        type: string
      - description: 创建时间
        in: query
        name: CreateTimeEnd
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeStart
        type: string
      - description: 更新时间
        in: query
        name: UpdateTimeEnd
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.UserProtocolResp'
                  type: array
              type: object
      summary: 用户协议列表-所有
      tags:
      - user_protocol-用户协议
...
swagger: "2.0"